# docs/api/openapi.yaml
...
openapi: "3.1.0"
info:
  title: sisys API
  description: AI 驱动的战略规划与决策智能平台 API
  version: "0.1.0"
  contact:
    name: sisys Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
paths:
...
  /auth/login:
    post:
      tags:
        - authentication
      summary: 用户登录
      description: 验证用户凭证并返回 JWT 令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "401":
          description: 无效凭证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "423":
          description: 账户已锁定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/search/evaluate:
    post:
      x-implemented: true
      tags:
        - search
      summary: 检索相关性评估
      description: |
        对检索结果执行 LLM-as-a-Judge 多维评估（相关性/完整性/时效性），
        综合评分 < 0.6 时标注"数据不足"并阻断摘要生成。
      security:
        - OAuth2: []
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
      responses:
        "200":
          description: 评估成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateResponse"
        "400":
          description: 请求参数校验失败
        "422":
          description: 检索结果不足被阻断
        "500":
          description: LLM 评估调用失败
components:
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
          description: 用户名
        password:
          type: string
          minLength: 1
          description: 密码
...
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          nullable: true
          description: Refresh token（可选）
        token_type:
          type: string
          default: bearer
          description: Token 类型
        expires_in:
          type: integer
          description: 过期时间（秒）
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      required:
        - id
        - username
        - roles
      properties:
        id:
          type: string
          description: 用户 UUID
        username:
          type: string
          description: 用户名
        roles:
          type: array
          items:
            type: string
          description: 角色列表

    # --------------------------------------------------------------------------
    # 角色相关 schemas
    # --------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: 错误详情

    # --------------------------------------------------------------------------
    # Embedding API schemas (Story 3-1a)
    # --------------------------------------------------------------------------
...
    EvaluateRequest:
      type: object
      required:
        - query_text
      properties:
        query_text:
          type: string
          minLength: 1
          description: 查询文本
        tenant_id:
          type: string
          nullable: true
          description: 租户 ID
      description: 检索相关性评估请求（系统内部执行真实检索后评估，不接受客户端传入检索结果）
    EvaluateResponse:
      type: object
      required:
        - overall_score
        - context_relevance
        - completeness
        - timeliness
        - context_relevance_reason
        - completeness_reason
        - timeliness_reason
        - should_block
        - block_reason
      properties:
        overall_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: 综合评分 ((context_relevance + completeness + timeliness) / 3)
        context_relevance:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: 上下文相关性评分
        completeness:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: 完整性评分
        timeliness:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: 时效性评分
        context_relevance_reason:
          type: string
          description: 相关性判断理由
        completeness_reason:
          type: string
          description: 完整性判断理由
        timeliness_reason:
          type: string
          description: 时效性判断理由
        should_block:
          type: boolean
          description: 是否阻断摘要生成（overall_score < 0.6 时为 true）
        block_reason:
          type: string
          nullable: true
          description: 阻断理由（should_block=True 时为"数据不足"，否则为 null）
      description: 检索相关性评估响应
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.example.com/authorize
          tokenUrl: https://auth.example.com/token
          scopes: {}
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token 认证